quickfix

+++ src/main/java/com/fuglkrig/server/Player.java
@@ -17,14 +17,14 @@ public class Player {
     int playerID;
     int hp;
     int skin;
-    int targetPosX;
-    int targetPosY;
-    int coordX;
-    int coordY;
+    float targetPosX;
+    float targetPosY;
+    float coordX;
+    float coordY;
     float xSpeed;
     float ySpeed;
-    int direction;
-    int speed;
+    float direction;
+    float speed;
     boolean alive;
 
     /**
@@ -203,21 +203,21 @@ public class Player {
     /**
      * @return the players target x position
      */
-    public int getTargetPosX() {
+    public float getTargetPosX() {
         return this.targetPosX;
     }
 
     /**
      * @return the players target y position
      */
-    public int getTargetPosY() {
+    public float getTargetPosY() {
         return this.targetPosY;
     }
 
     /**
      * @return the players current x position
      */
-    public int getCoordX() {
+    public float getCoordX() {
         return this.coordX;
     }
 
@@ -232,7 +232,7 @@ public class Player {
         this.coordY = y;
     }
 
-    public void setTargetPos(int x, int y){
+    public void setTargetPos(float x, float y){
         System.out.println("\nold pos: " + coordX + " " + coordY);
         System.out.println("target pos: " + targetPosX + " " + targetPosY);
 
@@ -255,21 +255,21 @@ public class Player {
     }
 
 
-    public int getCoordY() {
+    public float getCoordY() {
         return this.coordY;
     }
 
     /**
      * @return the players direction
      */
-    public int getDirection() {
+    public float getDirection() {
         return this.direction;
     }
 
     /**
      * @return the players current speed
      */
-    public int getSpeed() {
+    public float getSpeed() {
         return this.speed;
     }
